Use column name attribute to get session data
The call to the session data ResultSet was hard-coded to use the column
name "session_data", so if the database happened to store session data
in a column of a different name, the session was constantly re-created.

Changed the call to the ResultSet to use the data column name in the
corresponding attribute on $self and added the primary key to the
earlier "find".