Separate tests for "standard" and "custom" schemas
If the tests using the custom schema (Test::Custom) are run in the same
process as those using the default column names (Test::Schema), certain
ones fail. Running the two sets of tests separately causes all of them
to pass.

Looks like there's some strangeness in the way the schema classes are
loaded; down in the bowels the session object has a ResultSet that's
tied to the wrong Schema and ends up looking in the wrong column for the
session data.